Mix sound sources in 32-bit buffer, then clip to 16-bit for output. Fixes overflow in some situations, eg MegaRace on SB16.

+++ src/sound.h
@@ -1,6 +1,6 @@
 #include "timer.h"
 
-void sound_add_handler(void (*get_buffer)(int16_t *buffer, int len, void *p), void *p);
+void sound_add_handler(void (*get_buffer)(int32_t *buffer, int len, void *p), void *p);
 
 extern int sbtype;
